Refactoring code to supports real database with Sqlite local file

+++ GpsPoiApp.Tests/Unit/DistanceMeterTests.cs
@@ -12,7 +12,7 @@ public class DistanceMeterTests
         Point pointB = new Point(20, 10);
 
         // Act (action to be tested)
-        int distance = pointA.DistanceMeter(pointB);
+        double distance = pointA.DistanceMeter(pointB);
 
         // Assert (verification of the result)
         Assert.Equal(7, distance);
